@@ -1,11 +1,10 @@
 package com.app.gym.repository.model;
 
+import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MapsId;
 import jakarta.persistence.Table;
 
 /**
@@ -16,18 +15,19 @@ import jakarta.persistence.Table;
 public class ExerciseRoutine {
 
 	/** The id. */
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	@EmbeddedId
+	private ExerciseRoutinePK id;
 
 	/** The routine. */
 	@ManyToOne
 	@JoinColumn(name = "routine_id")
+	@MapsId("routineId")
 	private Routine routine;
 
 	/** The exercise. */
 	@ManyToOne
 	@JoinColumn(name = "exercise_id")
+	@MapsId("exerciseId")
 	private Exercise exercise;
 
 	/** The series. */
@@ -37,13 +37,24 @@ public class ExerciseRoutine {
 	private Integer repetitions;
 
 	/**
+	 * Instantiates a new exercise routine.
+	 */
+	public ExerciseRoutine() {
+		// Empty constructor
+	}
+
+	/**
+	 * Gets the id.
+	 *
 	 * @return the id
 	 */
-	public Long getId() {
+	public ExerciseRoutinePK getId() {
 		return id;
 	}
 
 	/**
+	 * Gets the routine.
+	 *
 	 * @return the routine
 	 */
 	public Routine getRoutine() {
@@ -51,6 +62,8 @@ public class ExerciseRoutine {
 	}
 
 	/**
+	 * Gets the exercise.
+	 *
 	 * @return the exercise
 	 */
 	public Exercise getExercise() {
@@ -58,6 +71,8 @@ public class ExerciseRoutine {
 	}
 
 	/**
+	 * Gets the series.
+	 *
 	 * @return the series
 	 */
 	public Integer getSeries() {
@@ -65,6 +80,8 @@ public class ExerciseRoutine {
 	}
 
 	/**
+	 * Gets the repetitions.
+	 *
 	 * @return the repetitions
 	 */
 	public Integer getRepetitions() {
